Switch the traffic.cambsvoyager.com feeds to use https:
The feeds on traffic.cambsvoyager.com switched to redirecting to https: on 2018-05-21.
Unfortunatly feedmaker didn't follow redirects so these feeds were lots. This in turn
caused all the web pages to be suppressed because they though there was no data to display.

This edit switches the feed to https:, ahead of more extensive work to add redirect
support to feedmaker (which in turn requies a Vertx upgrade, which in turn requires work
on dataserver to remove references to depricated methods).