Modify feedmaker to follow redirects when poling URLs
Vertx's http client wasn't capable of folowing http redirects until
vertx 3.4.0. With the upgrade to 3.5.2 this functionality can be added
which provides some defence against changes to the URLs it is scraping.

In particular, this would have prevented a long outage when
the main carpark capacity feed

  http://traffic.cambsvoyager.com/rssfeed.aspx?obj=64&filtertype=1

started responding with a redirect to the https version

  https://traffic.cambsvoyager.com/rssfeed.aspx?obj=64&filtertype=1