Correct (or add) "zone.map" key
This indicates (to the version of the zone map that uses the 'new' API)
whether or not a particular zone should be shown n the map. Without this,
most zones will be drawn as overlapping 'in' and 'out' areas.

The previous version of the map depended on list.json being manually
maintained so that it only has one instance of an 'in' or 'out' domain.
With the general addition of this flag, list.json can be automatically
maintained from the underlying individual configuration files.